added map and ability to add home location to map
